Refactoring of gpvolve. Remove GenotypePhenotypeMSM subclass.
