+++ gpvolve/__init__.py
@@ -1,16 +1,3 @@
 from .__version__ import __version__
 
-from . import simulate
-from . import markov
-from . import utils
-from . import check
-
-# from .markovmodel import GenotypePhenotypeMSM
-# from .visualization import *
-# from .utils import *
-# from .fitness import *
-# from .fixation import *
-# from .flux import *
-# from .paths import *
-# from .analysis import *
-# from .cluster import *
+from . import *
