+++ gpvolve/phenotype_to_fitness/__init__.py
@@ -1,0 +1,4 @@
+from .exponential import *
+from .linear import *
+from .sigmoidal import *
+from .step import *
