+++ gpvolve/utils/__init__.py
@@ -1,3 +1,4 @@
 
-from .flatten_neighbors import flatten_neighbors
-from .check_neighbor_connectivity import check_neighbor_connectivity
+from .flatten_neighbors import *
+from .check_neighbor_connectivity import *
+from .tools import *
